# apps/rust/openapi_output.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freefire",
    "description": "api gratis",
    "license": {
      "name": ""
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/komik/manhua": {
      "get": {
        "tags": [
          "komik"
        ],
        "operationId": "komik_manhua_slug",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (starts from 1)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Handles GET requests for the komik/manhua endpoint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManhuaResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/komik/manhwa": {
      "get": {
        "tags": [
          "komik"
        ],
        "operationId": "komik_manhwa_slug",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (starts from 1)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Handles GET requests for the komik/manhwa endpoint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManhwaResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ManhuaItem": {
        "type": "object",
        "required": [
          "title",
          "poster",
          "chapter",
          "score",
          "date",
          "type",
          "slug"
        ],
        "properties": {
          "chapter": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "poster": {
            "type": "string"
          },
          "score": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "ManhuaResponse": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManhuaItem"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "ManhwaItem": {
        "type": "object",
        "required": [
          "title",
          "poster",
          "chapter",
          "score",
          "date",
          "type",
          "slug"
        ],
        "properties": {
          "chapter": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "poster": {
            "type": "string"
          },
          "score": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "ManhwaResponse": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManhwaItem"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "required": [
          "current_page",
          "last_visible_page",
          "has_next_page",
          "has_previous_page"
        ],
        "properties": {
          "current_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "has_next_page": {
            "type": "boolean"
          },
          "has_previous_page": {
            "type": "boolean"
          },
          "last_visible_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "next_page": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "previous_page": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}